Revert "feat(assets-converter): Add /blog api"
