+++ .github/workflows/versions-upload.yaml
@@ -109,10 +109,8 @@ jobs:
     runs-on: ubuntu-latest
     if : >- 
       ${{ 
-        !cancelled() && 
-        !failure() &&
-        (needs.upload-website.outputs.deployment-url != '' ||
-        needs.upload-assets-converter.outputs.deployment-url != '')
+        needs.upload-website.outputs.deployment-url != '' || 
+        needs.upload-assets-converter.outputs.deployment-url != ''
       }}
     steps:
       - uses: actions/checkout@9c091bb21b7c1c1d1991bb908d89e4e9dddfe3e0 # v7.0.0
