+++ packages/website/src/type.ts
@@ -1,5 +1,6 @@
 export interface Bindings {
   TAGA3S_DEV_BUCKET: R2Bucket;
+  WORKERS_ENV: string;
   POLICY_AUD: string;
   TEAM_DOMAIN: string;
 }
